+++ src/Components/Main/ReviewForm.js
@@ -4,18 +4,18 @@ import Rating from '@mui/material/Rating';
 const ReviewForm = ({ review, onChange, onSubmit }) => {
   return (
     <div>
+                  <br></br>
       <form onSubmit={onSubmit}>
         <div>
           <Rating name="starRating" 
             size="small"
             defaultValue={2}
-            value={review.starRating}
+            value={review.starRating.toString()}
             onChange={onChange}
             />
         </div>
         <div>
-          <label>Review:</label>
-          <br />
+          <label>Review: </label>
           <input
             type="text"
             value={review.bodyText}